Update the name of kernel binary
The kernel build script has a variable to set the dist dir.
This could be quite convenient to use for development.
But this install a file name Image.gz-dtb, instead of kernel.
Update device.mk to use Image.gz-dtb instead of kernel.

Testing done:
Build the kernel using the build script, with DIST_DIR pointing to
device/mediatek/mt8183-kernel, and rebuild a bootimage that includes
the kernel built by the script.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>