[HACK] recovery: releasetools: bypass kernel compat test
When applying an OTA, the updater process fails while performing a
kernel compatibility check:

     Framework is incompatible with kernel version 4.19.52, compatible
     kernel versions are 4.4.107 4.4.107 4.4.107 4.9.84 4.9.84 4.9.84
     4.14.42 4.14.42 4.14.42

This check is only done if `compatibility.zip` is present
in the ota zip file.

The addition of `compatibility.zip` is controlled by
AddCompatibilityArchiveIfTrebleEnabled(), which reads the
`ro.treble.enabled` build property: if true, we copy
`compatibility.zip`.

This patch adds a custom release tools to avoid copying
`compatibility.zip` with the following flow:

  device_specific.FullOTA_InstallBegin()
    -> temporary setprop ro.treble.enabled false
  AddCompatibilityArchiveIfTrebleEnabled()
    -> does nothing, since property is false
  device_specific.FullOTA_InstallEnd()
    -> restore ro.treble.enabled to true

Note: this can be reverted when we switch to Q, as
Q is compatible with kernel 4.19.

Test:
    OUT_DIR=~/data/aosp-out DIST_DIR=~/data/aosp-dist-dir/ make dist
    # inspect ota_update.zip file, make sure compatibility.zip is
    # not present
    adb root
    adb reboot sideload
    cd ~/data/aosp-dist-dir/
    adb sideload mt8183-ota-*.zip
    # OTA applies successfully

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>