Rename the product and use external SoC definitions
The SoC definitions were defined here, in the product.
In order to make easier to start a new project for that SoC,
move all files specific to that SoC in common folder,
and include them from the product.
In addition, this rename the product with the name of the board.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>