kernel: commandline: remove root=/dev/X option
Currently, we are hard-coding the system partition number in the kernel
commandline.

Each time we update the partition scheme, we must also change this
commandline option, which is error prone.

Let the bootloader inspect the partition table and provide the
root=/dev/X argument to the kernel instead of hard-coding it here.

Bug: https://baylibre.atlassian.net/browse/RITA-65
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>