fstab: add soc specific fstab files
Currently, we have a different fstab for each product/variant.

This does not makes much sense, since most of the content of that fstab
is *soc* specific, not product specific.

This leads to code duplication and overall maintenance burden.

Add a soc-specific fstab file in here.

Note: If products have a need for custom partitions, they can:
 - create an additional fstab file which list those partitions
 - create a product specific init file which mounts the above partitions
   via mount_all

Bug: https://baylibre.atlassian.net/browse/RITA-67
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>