audio: allow product specific audio HAL configuration
For now, all products which belong to the same SoC family (mt8183,
mt8167) share the same audio hal configuration.

This is convenient for bring-up purposes, but each product will
eventually have hardware specific changes.

Add a mechanism so that a product can (optionally) override the audio
hal configuration file.

For a product to override the config we must add the following changes
to the device/mediatek/<product>/<product>.mk:

- Set TARGET_AUDIO_HAL_XML := <path to audio_configuration.xml>

Bug: RITA-114
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>