sepolicy: firmware: add custom firmware dt path
On MT8183 platforms, we use a custom firmware path in the device tree
so that products can override parts of the android firmware node.

This is set by the bootloader via the following kernel parameter:
  androidboot.android_dt_dir=/proc/device-tree/odm/firmware/android/

Add this custom path to the sysfs_dt_firmware_android context.

This commit fixes the following SELinux denials:

[    1.050367] audit: type=1400 audit(2724383810.736:4): avc:  denied  { read } for  pid=1 comm="init" name="compatible" dev="sysfs" ino=2145 scontext=u:r:init:s0 tcontext=u:object_r:sysfs:s0 tclass=file permissive=1
[    1.053532] audit: type=1400 audit(2724383810.736:5): avc:  denied  { open } for  pid=1 comm="init" path="/sys/firmware/devicetree/base/odm/firmware/android/compatible" dev="sysfs" ino=2145 scontext=u:r:init:s0 tcontext=u:object_r:sysfs:s0 tclass=file permissive=1

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>