Set to variables usd by ARMNN and NNAPI to select the driver
ARMNN driver use two system variables to implements getCapabilities.
If this is not set, NNAPI will privilege the use of ARMNN for Quantized
models whereas the CPU is faster.
This updates the init script to set these two variables.

Note: The value should be right but we may have to tweak them.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>