demo: add Tooploox CheckCounter demo app
During embedded world 2020, Tooploox developed a demo to recognize food
using an uvc camera on Opal.

Add their demo as part of the Opal build.

Note: by default Opal is in device USB mode, so the device tree needs to
be patched to enable host usb mode, so that we can use a camera.

Link:
https://www.mediatek.com/news-events/press-releases/mediateks-i500-aiot-platform-powers-tooplooxs-smart-self-checkout-application

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>