touchsreen: add calibration file to rotate 90° CCW
Currently, the touchscreen coordinates reported by the hardware are
rotated 90° Clock Wise (CW).

To compensate for that, we need to rotate the coordinates back 90°
Counter Clock Wise (CCW).

To do so, the formula is as following:
   x_new = (max_value - 1) - y
   y_new = x

Unfortunately, we cannot configure this via the standard .idc [1] way.

However, the framework can apply an affine transformation [2] on the
coordinates:

   newX = x * x_scale + y * x_ymix + x_offset;
   newY = x * y_xmix + y * y_scale + y_offset;

This gives us:

   newX = (max_value - 1) - y == x * x_scale + y * x_ymix + x_offset;
   <=> newX = (2047) - y == x * x_scale + y * x_ymix + x_offset
   If we tune the variables, we can have:
   <=> newX = (2047) - y == x * 0 + y * -1 + 2047
   <=> newX = (2047) - y ==  -y + 2047
   <=> true

   for newY, the logic is simlar so i did not detail it

Configure the calibration values in input-manager-state.xml, which is
read by Android on boot to rotate the touch coordinates.

[1] https://source.android.com/devices/input/input-device-configuration-files
[2] https://android.googlesource.com/platform/frameworks/native/+/refs/tags/android-10.0.0_r41/services/inputflinger/InputReaderBase.cpp#119

Test: adb shell dumpsys input | grep -A 6  'Affine Transformation:'

Reported-by: Andrew Perepech <andrew.perepech@mediatek.com>
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>