BoardConfig: add A/B OTA support
In the case of using A/B OTA support, the following actions should be
carried out:
- recovery partition must not be used
- cache partition must not be used
- super parition should have double the size as with legacy OTA.
- as a consequence of increasing super partition size, userdata
partition decreased as much. However, add to the size of userdata, the
size of recovery and cache partitions.

In order to comply with AOSP recommandations, substruct a 4MiB overhead
from BOARD_DB_DYNAMIC_PARTITIONS_SIZE as follows:

For A/B launch devices:
BBOARD_DB_DYNAMIC_PARTITIONS_SIZE = BOARD_SUPER_PARTITION_SIZE / 2 -
overhead

For non-A/B devices:
BBOARD_DB_DYNAMIC_PARTITIONS_SIZE = BOARD_SUPER_PARTITION_SIZE - overhead

In fact, Overhead is required in the computation to account for metadata,
alignments, and so on. A reasonable overhead is 4 MiB, it is possible to
pick a larger overhead as needed by the device.

RITA-372

Signed-off-by: Amjad Ouled-Ameur <aouledameur@baylibre.com>