Move userdata size flag to common/soc
From now on, the userdata partition will automatically take the remaining
eMMC space since it's the last partition in the partitioning table.

This is not only a choice, but a constraint enforced by the the flasher
(LK: little kernel), the last partition takes the remaining eMMC space
and must align 64KiB

Move BOARD_USERDATAIMAGE_PARTITION_SIZE to
common/soc/BoardConfigCommon.mk where the userdata will be automatically
computed, taking account of the contraints mentioned above.

RITA-402

Signed-off-by: Amjad Ouled-Ameur <aouledameur@baylibre.com>