Allocate remaining eMMC space to userdata partition
From now on, the userdata partition will automatically take the remaining
eMMC space since it's the last partition in the partitioning table.

This is not only a choice, but a constraint enforced by the the flasher
(LK: little kernel), the last partition takes the remaining eMMC space
and must align 64KiB

Compute BOARD_USERDATAIMAGE_PARTITION_SIZE autmoatically with
gen_partition_xml.py, and parse the result with 'awk' from
$(PRODUCT_OUT)/partitions.yaml

define MTK_PRODUCT_OUT with the same value as PRODUCT_OUT, the latter
is not reachable when BOARD_USERDATAIMAGE_PARTITION_SIZE is evaluated
the first time by Android Build System.

RITA-402

Signed-off-by: Amjad Ouled-Ameur <aouledameur@baylibre.com>