Provide support to build and run optee (TEE environment)
OP-TEE is a Trusted Execution Environment (TEE) designed as companion
to a non-secure Linux kernel running on Arm.

When OPTEE_ENABLE is set to true, all the necessary optee components
are built and added to the device.

When TEE_KEYMASTER_GATEKEEPER_ENABLE is set to true, we use
keymaster/gatekeeper from optee instead of the software implementation
from AOSP.

The xtest test suite is also added to the device with userdebug build.

Signed-off-by: Julien Masson <jmasson@baylibre.com>