common: install SystemUpdaterSample for A/B (userdebug)
To demonstrate A/B OTA updates, we use the SystemUpdaterSample sample
app which is provided by the AOSP (in //bootable/recovery)

Install it by default for userdebug builds so that we can easily
demonstrate/test A/B OTAs.

Note: end-user products should not use this. They should either use the
GMS services for updating or implement their own custom update service.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>