mt8183: audio: add the i2s audio configuration files
Add the audio config files to use i2s on the 40 pins header. Using
these 2 config files instead of the default ones will allow the
user to use the i2s1 2ch out and the i2s2 4ch in (limited to 2ch
in the android framework). In this case the jack will not be
operational anymore.

Signed-off-by: Nicolas <nbelin@baylibre.com>