mt8183|mt8167: fstab: Add new dynamic partitions to the vendor fstab
While we boot up properly with the dynamic partitions in only
the ramdisk fstab, Yongqin noted adb remount was failing recently.

On looking into the issue, I realized one issue is we don't have
the dynamic paritions listed in the vendor fstab which causes
some of the remount logic to not find the partitions.

This patch adds the dynamic partitions and unifies the fstab
file for all mtk soc so we have the same one for both the ramdisk
and vendor image.

Bug: RITA-468
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>