common: make device manifest SoC specific
The device manifest lists the HALS which are compatible with each
version. It also dictates the target level (AOSP major revision).

For android 11, we should be target-level=5.
This is possible on mt8183 based boards but not on mt8167 since the
graphic stack on mt8167 is too old.

To prepare to set target-level=5 for mt8183, remove the common manifest
file and use a SoC specific instead.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>