mt8183: inherit from optional cadence neural network HAL
According to i500 product brief [1], the i500 processor has a Tensilica
P6 which can be used for neural network processing.
This procesor is commonly called the VP6 here.

To be able to use the VP6, we need some cadence-restricted binaries.

Use an optional inheritance so that people who have access can use the
VP6, but for people who don't, we don't break the build.

[1] https://d86o2zu8ugzlg.cloudfront.net/mediatek-craft/documents/i500/MT8385-Product-Brief-1.pdf
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>