Support setting proj ID on external storage
After the migration from kernel 4.19 to kernel 5.4, few things have changed
on how app private directories (cache, obb and data) are created. The
former used SDCardFS which has been depecrated ; and no longer used on 5.4.
The latter needs some operations to be supported in order to create those
directories properly, one of those operations are setting the project ID
upon the created directory. This is currently not supported but is
necessary. In order to enable it, these two variables should be set
properly on the device.mk:

- PRODUCT_QUOTA_PROJID := 1
- PRODUCT_VENDOR_PROPERTIES += external_storage.projid.enabled=1

Signed-off-by: Amjad Ouled-Ameur <aouledameur@baylibre.com>