REVERTME: i500_pumpkin: update number of keepalive offload slots
This is to fix CTS tests related to wifi keepalive offload.
- testSocketKeepaliveUnprivileged
- testCreateTcpKeepalive
- testSocketKeepaliveLimitWifi

The tests call ConnectivityManager.createSocketKeepalive()
and then SocketKeepalive.start(), but the system returns
HARDWARE_ERROR instead of STARTED or UNSUPPORTED.
Investigation needed at wifi driver level.

By default, the number of keepalive offload slots is set to 3.
Override this value to zero to temporarily disable the feature.

Note: this patch is a workaround, until the HARDWARE_ERROR is fixed.
Signed-off-by: Matthieu Malis <mmalis@baylibre.com>