docs: getting-started: add note on power supply
Most of the time, the users connect the adb port of the board
directly to the usb port of the pc or the laptop.

If the adb port is also the power supply connector of the board,
the power might be too weak for enabling all peripherals of
the device.

Add a note in getting-started for the users to make sure
that power is high enough for the device to work completely.

Signed-off-by: Matthieu Malis <mmalis@baylibre.com>
(cherry picked from commit 7f5cd54f94290399a8da7c27c6fb43b1d3cab5f1)
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>