mt8183: vpud: use soong_namespace for vpud build
As of today, we have vpud binaries which use a `.mt8183` prefix but are
installed with a different name.
For example:
  name: "libvcodec_utility.mt8183",
Is installed as libvcodec_utility.so

This is forbidden by the Android build system for 12, as we can see in
this build error:

out/soong/.intermediates/device/mediatek/common/mt8183/binaries/vpud/libvcodec_utility.mt8183/android_vendor.31_arm64_armv8-a_cortex-a73_shared/libvcodec_utility.mt8183.so: error: DT_SONAME "libvcodec_utility.so" must be equal to the file name "libvcodec_u
tility.mt8183.so".

Fix it by using a soong_namespace, which allows us to have the same
module names for multiple modules.

Bug: RITA-727
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>