common: audio: Add USB audio support
Add the USB audio support on the platforms that can handle it.
This means:
- mt8183 i500-pumpkin platform
- mt8365 i350-evk and i350-pumpkin platforms

Eventhough audio is supported on the mt8167 i300 pumpkin platform,
the usb audio support has not been added on this platform as it
using USB audio causes kernel oops.

With this commit, you will be able to plug an USB headset and have
automatically audio playing on it. when you unplug your headset,
audio will play on your default audio device again.

Signed-off-by: Nicolas Belin <nbelin@baylibre.com>