mt8167|mt8183|mt8365: cleanup unknown packages for RITA SoCs
Remove unknown packages after enabling "enforce-product-packages-exist" macro for RITA SoCs mt8167, mt8183 and mt8365.
Only "libOpenCL.so" is required in "common/<SoC>/public.libraries.txt" ; Tested with OpenCL Z and Geekbench5.
"libOpenCL" could not be added to "PRODUCT_PACKAGES" in "<SoC>.mk" due to a Stub "src/mediatek/vendor/arm/libopencl-stub".
This Stub is required for other purposes, so we should not delete it ; for more info see commit message related to the Git Repo above.

Bug: RITA-820

Signed-off-by: Balsam CHIHI <bchihi@baylibre.com>