board: docs: update documentation
This commit removes all documentation related to the use
of "TARGET_USE_AB_SLOT". Since we dropped all noab build
support from AOSP and bootloaders scripts, specifying the
build type is not needed anymore. Henceforth, A/B build
is considered as the default and the only build supported
in all MediaTek projects.

BUG: RITA-990

Signed-off-by: fchiby <fchiby@baylibre.com>