mt8365: load kernel modules based on modules.load
The modalias_handling option defined in ueventd.rc was our solution to
"autoload" kernel modules based on modalias.
The downside of this solution is that ueventd will try to load/match
all the modalias found under /sys even if these drivers are already
built-in and loaded by the kernel.

Instead we can load the kernel modules based on modules.load
This file is auto-generated when we build the vendor image and
contains all the kernel modules needed for the corresponding device.

Android kernel modules infos:
https://source.android.com/docs/core/architecture/kernel/loadable-kernel-modules

Signed-off-by: Julien Masson <jmasson@baylibre.com>