device: rename, reorder and update comments
The default convention name in Android where we add files and modules
needed for a device is: device.mk
There is no reason to add suffix common to this file.

All the configs, properties, rc/xml files, modules/packages ... not
related to any feature folder for now have been moved to the end of
this file.

Several comments have been added or updated.

Signed-off-by: Julien Masson <jmasson@baylibre.com>