graphics: rework tree structure with: allocator, composer and EGL
When we modify something in the graphics stack, it usually concerns:
- allocator:
We support eith ARM gralloc or minigbm.
We can select one of the them with TARGET_ALLOCATOR_BACKEND.
By default we use ARM gralloc.

- composer:
We only use hwc3 composer.

- EGL:
We support either SwAngle or Mali.
We can select one of the them with TARGET_GPU_BACKEND.
By default we use Mali.

These 3 components have been moved to dedicated folder.

Signed-off-by: Julien Masson <jmasson@baylibre.com>